added *ADD ROLE* & *ADD EMPLOYEE*
added the functions called when the users selects the option to add an employee or a role. These functions include a prompt for the desired values to be added, as well as queries to pull ID data from the database, based on the response given in the prompt. not sure if its going to work yet.